feat: 应用商店兼容移动端
