+++ frontend/src/views/app-store/installed/index.vue
@@ -2,7 +2,7 @@
     <LayoutContent v-loading="loading || syncLoading" :title="activeName">
         <template #toolbar>
             <el-row :gutter="5">
-                <el-col :span="20">
+                <el-col :xs="24" :sm="24" :md="20" :lg="20" :xl="20">
                     <div>
                         <el-button
                             class="tag-button"
@@ -26,7 +26,7 @@
                         </div>
                     </div>
                 </el-col>
-                <el-col :span="4">
+                <el-col :xs="24" :sm="24" :md="4" :lg="4" :xl="4">
                     <div class="search-button">
                         <el-input
                             class="table-button"
@@ -68,7 +68,7 @@
                 >
                     <div class="install-card">
                         <el-card class="e-card">
-                            <el-row :gutter="24">
+                            <el-row :gutter="20">
                                 <el-col :xs="3" :sm="3" :md="3" :lg="4" :xl="4">
                                     <div class="icon">
                                         <el-avatar
@@ -78,7 +78,7 @@
                                         />
                                     </div>
                                 </el-col>
-                                <el-col :xs="21" :sm="21" :md="21" :lg="20" :xl="20">
+                                <el-col :xs="24" :sm="21" :md="21" :lg="20" :xl="20">
                                     <div class="a-detail">
                                         <div class="d-name">
                                             <el-button link type="info">
